move entry-modules logic into lazy-output.ts
